show_replaced_by_check_graphic as fixture
