+++ docs/gallery_tests/conftest.py
@@ -8,14 +8,18 @@
 Pytest fixtures for the gallery tests.
 
 """
-
+import collections
 import sys
 
+import matplotlib.pyplot as plt
 import pytest
 
 import iris
+import iris.plot as iplt
+import iris.quickplot as qplt
+from iris.tests import check_graphic
 
-from .gallerytest_util import gallery_path
+from .gallerytest_util import gallery_examples, gallery_path
 
 GALLERY_DIRECTORY = gallery_path()
 GALLERY_DIRECTORIES = [
@@ -52,3 +56,28 @@ def iris_future_defaults():
         del default_future_kwargs[dead_option]
     with iris.FUTURE.context(**default_future_kwargs):
         yield
+
+
+@pytest.fixture(params=gallery_examples())
+def show_replaced_by_check_graphic(request):
+    """
+    Creates a fixture which, for the gallery examples specified by params, is
+    used to replace the functionality of matplotlib.pyplot.show with a function
+    which calls the check_graphic function (iris.tests.check_graphic).
+
+    Yields the example name so it can be imported in the test.
+
+    """
+    test_id = f"gallery_tests.test_{request.param}"
+    assertion_counts = collections.defaultdict(int)
+
+    def replacement_show():
+        # form a closure on test_case and tolerance
+        unique_id = f"{test_id}.{assertion_counts[test_id]}"
+        assertion_counts[test_id] += 1
+        check_graphic(unique_id)
+
+    orig_show = plt.show
+    plt.show = iplt.show = qplt.show = replacement_show
+    yield request.param
+    plt.show = iplt.show = qplt.show = orig_show
